+++ packages/gov-ui-kit/src/components/input/inputText/index.ts
@@ -1,1 +1,2 @@
-export { InputText, type IInputTextProps } from './inputText';
+export { InputText } from './inputText';
+export type { IInputTextProps } from './inputText.api';