+++ packages/gov-ui-kit/CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 -   Handle size property on `Progress` component
 -   Implement `InputTime` component
 
+### Changed
+
+-   Update minor and patch versions of dependencies
+
 ## [1.0.9] - 2024-01-23
 
 ### Fixed