+++ packages/gov-ui-kit/CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 -   Update minor and patch versions of dependencies
+-   Update `husky` to v9
 
 ## [1.0.9] - 2024-01-23