+++ packages/gov-ui-kit/CHANGELOG.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 -   Add `wrapperClassName` property to `InputContainer` component to customise the input wrapper
 -   Update `InputContainer` props to accept any HTML div property and support textarea elements
 
+### Fixed
+
+-   Fix styling conflict in `InputText` introduced by Addon element
+
 ## [1.0.9] - 2024-01-23
 
 ### Fixed